Module dumbmath::vector

Structs

Vec3f
Vector with three f32 components

Constants

ZERO_3F
Vec3f(0.0, 0.0, 0.0)

Traits

CastF32
Convert a numeric type to an f32

Functions

cross
Cross product of two Vec3f inputs

distance3
Distance between two points

dot3
Inner product of two Vec3f inputs

lerp3
Linearly interpolate between two points by the factor t. When t is zero the result is p0, and when t is one the result is p1. The range of t is not clamped.

vec3f
Create a Vec3f from x, y, and z inputs