Function dumbmath::vector::vec3f

pub fn vec3f<X: CastF32, Y: CastF32, Z: CastF32>(x: X, y: Y, z: Z) -> Vec3f

Create a Vec3f from x, y, and z inputs

This is a convenience function that provides a little more flexibility than Vec3f::new in that it will happily take numbers that aren't f32 (including a mix of different types for each component). Vec3f is such a common type that it seems reasonable to provide a little extra ease of use.