Struct dumbmath::vector::Vec3f

pub struct Vec3f {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Vector with three f32 components

Fields

x

y

z

Methods

impl Vec3f

fn new(x: f32, y: f32, z: f32) -> Vec3f

Create a Vec3f from three components

fn from_scalar<T: CastF32>(s: T) -> Vec3f

Create a Vec3f with all components set to the same scalar

fn magnitude_squared(&self) -> f32

Squared length of the vector

fn magnitude(&self) -> f32

Length of the vector

fn normalized(&self) -> Option<Vec3f>

Create a normalized copy, or None if the magnitude is zero

fn project_onto(&self, v: Vec3f) -> Option<Vec3f>

Projection of self into v, or None if the magnitude of b is zero.