Struct dumbmath::segment::Segment3f

pub struct Segment3f {
    pub start: Vec3f,
    pub end: Vec3f,
}

Line segment between two points

Fields

start

end

Methods

impl Segment3f

fn new(start: &Vec3f, end: &Vec3f) -> Segment3f

fn length(&self) -> f32

fn closest_point_to_point(&self, point: &Vec3f) -> (f32, Vec3f)

Find the point on the segment closest to the input point. The return value contains both the parametric and actual location of the closest point.

Adapted from "Real-Time Collision Detection" by Christer Ericson, published by Morgan Kaufmann Publishers, Copyright 2005 Elsevier Inc